+++ Neoforge/src/main/java/com/tristankechlo/healthcommand/NeoforgePlatformHelper.java
@@ -1,0 +1,15 @@
+package com.tristankechlo.healthcommand;
+
+import com.tristankechlo.healthcommand.platform.IPlatformHelper;
+import net.neoforged.fml.loading.FMLPaths;
+
+import java.nio.file.Path;
+
+public class NeoforgePlatformHelper implements IPlatformHelper {
+
+    @Override
+    public Path getConfigDirectory() {
+        return FMLPaths.CONFIGDIR.get();
+    }
+
+}