+++ Changelog.md
@@ -3,6 +3,7 @@
 ### Version 1.20.4 - 1.3
 
 - port to 1.20.4
+- add neoforge support
 
 ### Version 1.20.1 - 1.3